solucionado lo de la escala del terreno y todo eso

+++ trunk/src/codigo/TP taller/TP taller/Parser/yaml/ParserYaml.h
@@ -16,8 +16,8 @@
 using namespace std;
 
 struct EscenarioParseado{
-	int altoPx;
-    int anchoPx;
+	double altoPx;
+    double anchoPx;
     int altoU;
     int anchoU;
     int nivelAgua;
